implement RGB to RGBA conversion in ARM64 NEON assembler
This conversion used to take about 3s (for 4960x7016 RGB pixels).
With the new code it takes about 300ms.

More potential for improvement: we could run this code while reading
pixels via USB.

I haven’t looked at the USB timing in detail, but my guess is that we
could squeeze in this post-processing into the time between requesting
data from the device and receiving the data from the kernel.

If that doesn’t work out, we could parallelize and post-process the
previous buffer while reading the current buffer.

Note that we need to use the WORD instruction because the Go assembler
is lacking support for the NEON instructions, see
https://github.com/golang/go/issues/7300

related to issue #7