Create /perm/scans/<sub> in `UpdateFromDir`
As it stands, after a new user signs into Google the index handler
returns an HTTP 500 with `open /perm/scans/<id>: no such file or
directory`. As far as I can tell this is coming from `Queue.Scans()`,
and happens because nothing is creating `/perm/scans/<id>`.
`Queue.AddJob()` will create it eventually but obviously without the UI
working it's quite hard to trigger a scan job to do this.

Therefore, create the per-user scans directory (and any parents) when
the users table is updated.